Added ECS; Added various components; Added WeakRef; Added UUID; Added SceneManager; Properly implemented HierarchyTree; Improved ModelLoader; Fixed incorrect Renderer calls;

+++ editor/src/ui/editorWindow.h
@@ -14,7 +14,7 @@ namespace emerald {
 		void onImGuiRender();
 		void destroy();
 
-		glm::vec2 getSceneViewportSize() const;
+		glm::ivec2 getSceneViewportSize() const;
 
 	private:
 		Ref<Texture> m_icon;
